+++ gameserver/src/main/java/com/lhf/game/map/Land.java
@@ -4,6 +4,7 @@ import java.io.FileNotFoundException;
 import java.io.Serializable;
 import java.util.Collection;
 import java.util.Collections;
+import java.util.EnumMap;
 import java.util.Map;
 import java.util.Objects;
 import java.util.Optional;
@@ -16,6 +17,7 @@ import java.util.logging.Level;
 import com.google.common.base.Function;
 import com.lhf.game.AffectableEntity;
 import com.lhf.game.CreatureContainer;
+import com.lhf.game.TickType;
 import com.lhf.game.creature.ICreature;
 import com.lhf.game.creature.conversation.ConversationManager;
 import com.lhf.game.creature.intelligence.AIRunner;
@@ -23,10 +25,23 @@ import com.lhf.game.creature.statblock.StatblockManager;
 import com.lhf.game.map.Area.AreaBuilder;
 import com.lhf.game.map.Area.AreaBuilder.AreaBuilderID;
 import com.lhf.game.map.Atlas.AtlasMappingItem;
+import com.lhf.game.map.Atlas.TargetedTester;
+import com.lhf.game.map.commandHandlers.LandSeeHandler;
+import com.lhf.game.map.commandHandlers.LandShoutHandler;
+import com.lhf.messages.Command;
 import com.lhf.messages.CommandChainHandler;
+import com.lhf.messages.CommandContext;
+import com.lhf.messages.CommandContext.Reply;
 import com.lhf.messages.GameEventProcessor;
 import com.lhf.messages.ITickEvent;
+import com.lhf.messages.events.BadGoEvent;
+import com.lhf.messages.events.BadGoEvent.BadGoType;
+import com.lhf.messages.events.BadMessageEvent;
+import com.lhf.messages.events.BadMessageEvent.BadMessageType;
 import com.lhf.messages.events.GameEvent;
+import com.lhf.messages.events.TickEvent;
+import com.lhf.messages.in.AMessageType;
+import com.lhf.messages.in.GoMessage;
 import com.lhf.server.client.user.UserID;
 
 public interface Land extends CreatureContainer, CommandChainHandler, AffectableEntity<DungeonEffect> {
@@ -235,6 +250,106 @@ public interface Land extends CreatureContainer, CommandChainHandler, Affectable
                 .filter(area -> area != null)
                 .forEach(area -> creatures.addAll(area.getCreatures()));
         return Collections.unmodifiableSet(creatures);
+    }
+
+    public interface LandCommandHandler extends CommandHandler {
+
+        static final EnumMap<AMessageType, CommandHandler> landCommandHandlers = new EnumMap<>(Map.of(
+                AMessageType.GO, new LandGoHandler(),
+                AMessageType.SEE, new LandSeeHandler(),
+                AMessageType.SHOUT, new LandShoutHandler()));
+
+        @Override
+        public default boolean isEnabled(CommandContext ctx) {
+            if (ctx == null) {
+                return false;
+            }
+            final ICreature creature = ctx.getCreature();
+            if (creature == null || !creature.isAlive()) {
+                return false;
+            }
+            final Area area = ctx.getArea();
+            if (area == null) {
+                return false;
+            }
+            final Land land = ctx.getLand();
+            return land != null;
+        }
+
+        @Override
+        public default CommandChainHandler getChainHandler(CommandContext ctx) {
+            return ctx.getLand();
+        }
+    }
+
+    public static class LandGoHandler implements LandCommandHandler {
+        private final static String helpString = "\"go [direction]\" Move in the desired direction, if that direction exists.  Like \"go east\"";
+
+        @Override
+        public AMessageType getHandleType() {
+            return AMessageType.GO;
+        }
+
+        @Override
+        public Optional<String> getHelp(CommandContext ctx) {
+            return Optional.of(LandGoHandler.helpString);
+        }
+
+        @Override
+        public Reply handleCommand(CommandContext ctx, Command cmd) {
+            if (cmd != null && cmd.getType() == this.getHandleType()) {
+                final GoMessage goMessage = new GoMessage(cmd);
+                final Land land = ctx.getLand();
+                if (ctx.getCreature() == null) {
+                    ctx.receive(BadMessageEvent.getBuilder().setBadMessageType(BadMessageType.CREATURES_ONLY)
+                            .setHelps(ctx.getHelps()).setCommand(cmd).Build());
+                    return ctx.handled();
+                }
+                Directions toGo = goMessage.getDirection();
+                if (ctx.getArea() == null) {
+                    ctx.receive(BadGoEvent.getBuilder().setSubType(BadGoType.NO_ROOM).setAttempted(toGo).Build());
+                    return ctx.handled();
+                }
+                Area presentRoom = ctx.getArea();
+                final AtlasMappingItem<Area, UUID> mappingItem = land.getAtlas()
+                        .getAtlasMappingItem(presentRoom.getUuid());
+                if (mappingItem != null) {
+                    Map<Directions, TargetedTester<UUID>> exits = mappingItem.getDirections();
+                    if (exits == null || exits.size() == 0
+                            || !exits.containsKey(toGo)
+                            || exits.get(toGo) == null) {
+                        ctx.receive(BadGoEvent.getBuilder().setSubType(BadGoType.DNE).setAttempted(toGo).Build());
+                        return ctx.handled();
+                    }
+                    TargetedTester<UUID> doorway = exits.get(toGo);
+                    final Area nextRoom = land.getAtlas().getAtlasMember(doorway.getTargetId());
+                    if (nextRoom == null) {
+                        ctx.receive(BadGoEvent.getBuilder().setSubType(BadGoType.DNE).setAttempted(toGo)
+                                .setAvailable(exits.keySet()).Build());
+                        return ctx.handled();
+                    }
+                    TraversalTester tester = doorway.getPredicate();
+                    if (tester != null && !tester.testTraversal(ctx.getCreature(), toGo, presentRoom, presentRoom)) {
+                        ctx.receive(BadGoEvent.getBuilder().setSubType(BadGoType.BLOCKED).setAttempted(toGo)
+                                .setAvailable(exits.keySet()).Build());
+                        return ctx.handled();
+                    }
+
+                    if (presentRoom.removeCreature(ctx.getCreature(), toGo)) {
+                        ICreature.eventAccepter.accept(ctx.getCreature(),
+                                TickEvent.getBuilder().setTickType(TickType.ROOM).Build());
+                        nextRoom.addCreature(ctx.getCreature());
+                        return ctx.handled();
+                    }
+                } else {
+                    ctx.receive(BadGoEvent.getBuilder().setSubType(BadGoType.NO_ROOM)
+                            .setAttempted(goMessage.getDirection()).Build());
+                    return ctx.handled();
+                }
+            }
+            return ctx.failhandle();
+        }
+
     }
 
     @Override
