+++ gameserver/src/main/java/com/lhf/game/map/commandHandlers/RestingGoHandler.java
@@ -4,9 +4,8 @@ import java.util.EnumSet;
 import java.util.Optional;
 
 import com.lhf.game.map.Directions;
-import com.lhf.game.map.RestArea;
-import com.lhf.game.map.SubArea;
 import com.lhf.game.map.RestArea.RestingCommandHandler;
+import com.lhf.game.map.SubArea;
 import com.lhf.messages.Command;
 import com.lhf.messages.CommandContext;
 import com.lhf.messages.CommandContext.Reply;
