+++ gameserver/src/main/java/com/lhf/game/map/commandHandlers/RestingRestHandler.java
@@ -2,9 +2,8 @@ package com.lhf.game.map.commandHandlers;
 
 import java.util.Optional;
 
-import com.lhf.game.map.RestArea;
-import com.lhf.game.map.SubArea;
 import com.lhf.game.map.RestArea.RestingCommandHandler;
+import com.lhf.game.map.SubArea;
 import com.lhf.messages.Command;
 import com.lhf.messages.CommandContext;
 import com.lhf.messages.CommandContext.Reply;
